pyproject.toml
[project]
name = "cloud-sentinel"
version = "0.1.0"
description = "application monitor"
authors = [
    {name = "Lucas Sousa"}
]
readme = "README.md"
requires-python = "^3.9"
dependencies = [
    "fastapi (>=0.115.14,<0.116.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "fastapi-admin (>=1.0.4,<2.0.0)",
    "prometheus-client (>=0.22.1,<0.23.0)",
    "psutil (>=7.0.0,<8.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "ping3 (>=4.0.8,<5.0.0)",
    "tortoise-orm (>=0.25.1,<0.26.0)",
    "redis (>=6.2.0,<7.0.0)",
    "aioredis (==2.0.1)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
